Type Alias jrsonnet_gcmodule::Cc

pub type Cc<T> = RawCc<T, ObjectSpace>;

A single-threaded reference-counting pointer that integrates with cyclic garbage collection.

See module level documentation for more details.

Cc is not thread-safe. It does not implement Send or Sync:

use std::ops::Deref;
use gcmodule::Cc;
let cc = Cc::new(5);
std::thread::spawn(move || {
    println!("{}", cc.deref());
});

Aliased Type

struct Cc<T>(/* private fields */);

Implementations

impl<T: Trace> Cc<T>

pub fn new(value: T) -> Cc<T>

Constructs a new Cc<T> in a thread-local storage.

To collect cycles, use collect_thread_cycles.

impl<T: Trace + Clone> Cc<T>

pub fn update_with(&mut self, update_func: impl FnMut(&mut T))

Update the value T in a copy-on-write way.

If the ref count is 1, the value is updated in-place. Otherwise a new Cc<T> will be created.

impl<T: Trace, O: AbstractObjectSpace> RawCc<T, O>

pub fn into_dyn(self) -> RawCc<dyn Trace, O>

Convert to RawCc<dyn Trace>.

impl<T: ?Sized, O: AbstractObjectSpace> RawCc<T, O>

pub fn downgrade(&self) -> RawWeak<T, O>

Obtains a “weak reference”, a non-owning pointer.

pub fn strong_count(&self) -> usize

Gets the reference count not considering weak references.

pub fn ptr_eq(this: &Self, other: &Self) -> bool

Returns true if the two Ccs point to the same allocation

impl<T: ?Sized, O: AbstractObjectSpace> RawCc<T, O>

pub fn trace(&self, tracer: &mut Tracer<'_>)

trace without T: Trace bound.

Useful for structures with Cc<T> fields where T does not implement Trace. For example, struct S(Cc<Box<dyn MyTrait>>). To implement Trace for S, it can use Cc::trace(&self.0, tracer).

pub fn weak_count(&self) -> usize

Get the weak (non-owning) reference count.