Type Alias jrsonnet_gcmodule::ThreadedCc

pub type ThreadedCc<T> = RawCc<T, ThreadedObjectSpace>;

A multi-thread reference-counting pointer that integrates with cyclic garbage collection.

ThreadedCc is similar to Cc. It works with multi-thread but is significantly slower than Cc.

To construct a ThreadedCc, use ThreadedObjectSpace::create.

Aliased Type

struct ThreadedCc<T>(/* private fields */);

Implementations

impl<T: ?Sized> ThreadedCc<T>

pub fn borrow(&self) -> ThreadedCcRef<'_, T>

Immutably borrows the wrapped value.

The borrow lasts until the returned value exits scope.

impl<T: Trace, O: AbstractObjectSpace> RawCc<T, O>

pub fn into_dyn(self) -> RawCc<dyn Trace, O>

Convert to RawCc<dyn Trace>.

impl<T: ?Sized, O: AbstractObjectSpace> RawCc<T, O>

pub fn downgrade(&self) -> RawWeak<T, O>

Obtains a “weak reference”, a non-owning pointer.

pub fn strong_count(&self) -> usize

Gets the reference count not considering weak references.

pub fn ptr_eq(this: &Self, other: &Self) -> bool

Returns true if the two Ccs point to the same allocation

impl<T: ?Sized, O: AbstractObjectSpace> RawCc<T, O>

pub fn trace(&self, tracer: &mut Tracer<'_>)

trace without T: Trace bound.

Useful for structures with Cc<T> fields where T does not implement Trace. For example, struct S(Cc<Box<dyn MyTrait>>). To implement Trace for S, it can use Cc::trace(&self.0, tracer).

pub fn weak_count(&self) -> usize

Get the weak (non-owning) reference count.