Type Alias jrsonnet_gcmodule::Weak

pub type Weak<T> = RawWeak<T, ObjectSpace>;

Weak reference of Cc.

Aliased Type

struct Weak<T>(/* private fields */);

Implementations

impl<T: ?Sized, O: AbstractObjectSpace> RawWeak<T, O>

pub fn upgrade(&self) -> Option<RawCc<T, O>>

Attempts to obtain a “strong reference”.

Returns None if the value has already been dropped.

pub fn strong_count(&self) -> usize

Gets the reference count not considering weak references.

pub fn weak_count(&self) -> usize

Get the weak (non-owning) reference count.

pub fn ptr_eq(this: &Self, other: &Self) -> bool

Returns true if the two Weaks point to the same allocation